+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+### Fixes and improvements
+
+* (client) avoid content corruption when `file` output string contains non utf-8 or non supported characters
+
 ## [v1.4.2](https://github.com/OpenNMT/nmt-wizard/releases/tag/v1.4.2) (2018-12-17)
 
 ### Fixes and improvements