+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixes and improvements
 
 * (client) avoid content corruption when `file` output string contains non utf-8 or non supported characters
+* (client)  improve task list to add model and image tag information
 
 ## [v1.4.2](https://github.com/OpenNMT/nmt-wizard/releases/tag/v1.4.2) (2018-12-17)